Configure logging in App.xaml.cs
Puts logging configuration in shared code and enables logging for the Android head.
