+++ Microsoft.Toolkit.Uwp.SampleApp.Wasm/Program.cs
@@ -12,44 +12,8 @@ namespace Microsoft.Toolkit.Uwp.SampleApp.Wasm
 
 		static void Main(string[] args)
 		{
-#if DEBUG 
-            ConfigureFilters(LogExtensionPoint.AmbientLoggerFactory);
-#endif
 
 			Application.Start(e => new App());
-		}
-         
-		static void ConfigureFilters(ILoggerFactory factory) 
-		{
-			factory
-				.WithFilter(new FilterLoggerSettings 
-					{
-						{ "Uno", LogLevel.Warning },
-						{ "Windows", LogLevel.Warning },
-						{ "SampleControl.Presentation", LogLevel.Debug },
-						//{ "Windows.UI.Xaml.Controls.VirtualizingPanelLayout", LogLevel.Debug },
-
-					// Generic Xaml events
-					// { "Windows.UI.Xaml", LogLevel.Debug },
-
-					// { "Uno.UI.Controls.AsyncValuePresenter", LogLevel.Debug },
-					// { "Uno.UI.Controls.IfDataContext", LogLevel.Debug },
-
-					// Layouter specific messages
-					// { "Windows.UI.Xaml.Controls", LogLevel.Debug },
-					// { "Windows.UI.Xaml.Controls.Layouter", LogLevel.Debug },
-					// { "Windows.UI.Xaml.Controls.Panel", LogLevel.Debug },
-
-					// Binding related messages
-					// { "Windows.UI.Xaml.Data", LogLevel.Debug },
-					// { "Windows.UI.Xaml.DependencyObjectStore", LogLevel.Debug },
-					// { "Uno.UI.DataBinding.BindingPropertyHelper", LogLevel.Debug },
-
-					//  Binder memory references tracking
-					// { "ReferenceHolder", LogLevel.Debug },
-				}
-				)
-				.AddConsole(LogLevel.Debug);
 		}
 	}
 }
