+++ Microsoft.Toolkit.Uwp.SampleApp.iOS/Main.cs
@@ -9,48 +9,11 @@ namespace Microsoft.Toolkit.Uwp.SampleApp.iOS
 		// This is the main entry point of the application.
 		static void Main(string[] args)
 		{
-#if DEBUG
-			ConfigureFilters(LogExtensionPoint.AmbientLoggerFactory);
-#endif
-
 			// if you want to use a different Application Delegate class from "AppDelegate"
 			// you can specify it here.
 			UIApplication.Main(args, null, typeof(App)); 
 		}
 
-		static void ConfigureFilters(ILoggerFactory factory)
-		{
-			factory
-				.WithFilter(new FilterLoggerSettings
-					{
-						{ "Uno", LogLevel.Warning },
-						{ "Windows", LogLevel.Warning },
-						{ "SampleControl.Presentation", LogLevel.Debug },
-
-					// Generic Xaml events
-					// { "Windows.UI.Xaml", LogLevel.Debug },
-
-					// { "Uno.UI.Controls.AsyncValuePresenter", LogLevel.Debug },
-					// { "Uno.UI.Controls.IfDataContext", LogLevel.Debug },
-					 { "Windows.UI.Xaml.FrameworkElement", LogLevel.Debug },
-					 { "Windows.UI.Xaml.UIElement", LogLevel.Debug },
-					 { "Windows.UI.Xaml.Controls.SinglelineTextBoxView", LogLevel.Debug },
-
-					// Layouter specific messages
-					// { "Windows.UI.Xaml.Controls", LogLevel.Debug },
-					//{ "Windows.UI.Xaml.Controls.Layouter", LogLevel.Debug },
-					//{ "Windows.UI.Xaml.Controls.Panel", LogLevel.Debug },
-
-					// Binding related messages
-					 // { "Windows.UI.Xaml.Data", LogLevel.Debug },
-					//{ "Windows.UI.Xaml.DependencyObjectStore", LogLevel.Debug },
-					 { "Uno.UI.DataBinding.BindingPropertyHelper", LogLevel.Debug },
-
-					//  Binder memory references tracking
-					// { "ReferenceHolder", LogLevel.Debug },
-				}
-				)
-				.AddDebug(LogLevel.Debug);
-		}
+		
 	}
 }
